refactors TimeAgoFormat function to its own file
