+++ frontend/src/components/leaderboards/CompetitionCard.tsx
@@ -132,7 +132,7 @@ export function CompetitionCard({
         textArea.select();
         try {
           // sonar error expected, its a fall back for older browsers
-          const success = document.execCommand("copy"); // eslint-disable-line @typescript-eslint/no-deprecated
+          const success = document.execCommand("copy");  
           if (!success) throw new Error("execCommand copy failed");
         } finally {
           textArea.remove();
